Use create job command instead of run command to collect metrics
curl-metrics test tries to run a curl command to collect the metrics and
it is intended to be short lived. This is typical use case for jobs rather
than pods. This commit modifies the test to run job by using kubectl create job
command. Besides, kubectl run is not recommended.