Use create jobs command instead of run pods in e2e testing (#445)
* Use create job command instead of run command to collect metrics

curl-metrics test tries to run a curl command to collect the metrics and
it is intended to be short lived. This is typical use case for jobs rather
than pods. This commit modifies the test to run job by using kubectl create job
command. Besides, kubectl run is not recommended.

* Use unprivileged nginx image instead of default nginx

Default nginx image requires privileged access in the pod which is
unwanted and not recommended even in testing environments.

This commit switches to nginxinc/nginx-unprivileged that installs nginx
in an unprivileged way. This commit also updates the nginx version to 1.27
instead of the very old 1.14.2