Fix issue with rolling update when some existing replicas are unhealthy (#488)
* Fix issue with rolling update when some existing replicas are unhealthy

The bug is as follows:
* assume replicas readiness is [NOT_READY, READY]
* update is triggered, with MaxUnavailable=1
* update starts updating the replica-1, causing complete downtime,
  while the desired behavior would be to wait for replica-0 to become ready

The fix is to reduce partition less aggressively - taking into account pods
that are not ready.

Note that the fix was tested on a cluster with MaxUnavailableStatefulSet feature disabled
(I don't have access to a cluster with this feature enabled, maybe the issue is less impactful there,
however arguably still beneficial since the lws considers readiness of leader+workers and not only leader).

Testing:
* integration tests

* Address feedback from pr

https://github.com/kubernetes-sigs/lws/pull/488

* Simplify the partition calculation logic

We make two simplifications:
* for unavailable we count only replicas that are <= "minAllowedPartition"
* we have a loop that can push the result partition down as long as the replicas are
unavailable, that would prevent the update getting stuck scenario

This makes the overall logic much simpler and, even though sometimes does not reduce the
aggressively enough - it'll be much better for long-term maintainablility of the code.

Testing:
* make test-intergation